Included textacular to do full text search Added search page at frontend
