+++ Gemfile
@@ -15,8 +15,9 @@ gem "thin"
 gem "bcrypt"
 gem "redis"
 gem "sidekiq"
-gem  "dotenv"
+gem "dotenv"
 gem "bunny"
+gem "textacular"
 
 # db
 gem "pg"
@@ -27,4 +28,3 @@ gem "byebug"
 gem "faker"
 gem "activerecord-import"
 gem 'newrelic_rpm'
-
