+++ client/src/components/general/navbar.js
@@ -16,9 +16,6 @@ import AccountCircle from '@material-ui/icons/AccountCircle';
 import MailIcon from '@material-ui/icons/Mail';
 import NotificationsIcon from '@material-ui/icons/Notifications';
 import MoreIcon from '@material-ui/icons/MoreVert';
-import Button from '@material-ui/core/Button';
-
-import { withRouter } from 'react-router-dom';
 
 import { nanoAPI } from '../../nanoAPI'
 import { sessionHelper } from '../../helpers/session'
